requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.1.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiofiles=23.2.1=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.11.13=pypi_0
aiosignal=1.3.2=pypi_0
altair=5.5.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.8.0=pypi_0
asttokens=3.0.0=pypi_0
async-timeout=5.0.1=pypi_0
attrs=25.2.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.13.3=pypi_0
bleach=6.2.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
certifi=2025.1.31=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
cmake=3.31.6=pypi_0
contourpy=1.3.1=pypi_0
customtkinter=5.2.2=pypi_0
cycler=0.12.1=pypi_0
darkdetect=0.8.0=pypi_0
dearpygui=2.0.0=pypi_0
decorator=5.2.1=pypi_0
defusedxml=0.7.1=pypi_0
docopt=0.6.2=pypi_0
exceptiongroup=1.2.2=pypi_0
executing=2.2.0=pypi_0
fastapi=0.115.11=pypi_0
fastjsonschema=2.21.1=pypi_0
ffmpy=0.5.0=pypi_0
filelock=3.17.0=pypi_0
fonttools=4.56.0=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2025.3.0=pypi_0
glfw=2.2.0=pypi_0
gradio=3.36.1=pypi_0
gradio-client=0.10.0=pypi_0
groovy=0.1.2=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.29.3=pypi_0
idna=3.10=pypi_0
imageio=2.37.0=pypi_0
imageio-ffmpeg=0.4.3=pypi_0
imgui=2.0.0=pypi_0
importlib-resources=6.5.2=pypi_0
ipython=8.12.3=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
kiwisolver=1.4.8=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=13.2.0=hc0a3c3a_7
libuuid=1.41.5=h5eee18b_0
linkify-it-py=2.0.3=pypi_0
lit=18.1.8=pypi_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.10.1=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdit-py-plugins=0.3.3=pypi_0
mdurl=0.1.2=pypi_0
mistune=3.1.3=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
narwhals=1.30.0=pypi_0
nbclient=0.10.2=pypi_0
nbconvert=7.16.6=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.4.2=pypi_0
ninja=1.11.1.3=pypi_0
numpy=1.23.5=pypi_0
opencv-python=4.11.0.86=pypi_0
openssl=1.1.1w=h7f8727e_0
orjson=3.10.15=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.5.0=pypi_0
pip=25.0=py310h06a4308_0
pipreqs=0.5.0=pypi_0
platformdirs=4.3.7=pypi_0
prompt-toolkit=3.0.50=pypi_0
propcache=0.3.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pydub=0.25.1=pypi_0
pygments=2.19.1=pypi_0
pyopengl=3.1.5=pypi_0
pyparsing=3.2.1=pypi_0
pyqt5=5.15.9=pypi_0
pyqt5-plugins=5.15.9.2.3=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.17.0=pypi_0
pyqt5-tools=5.15.9.3.3=pypi_0
pyspng=0.1.2=pypi_0
python=3.10.6=haa1d7c7_1
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-multipart=0.0.20=pypi_0
pytz=2025.1=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.3.0=pypi_0
qt5-applications=5.15.2.2.3=pypi_0
qt5-tools=5.15.2.1.3=pypi_0
readline=8.2=h5eee18b_0
referencing=0.36.2=pypi_0
requests=2.32.3=pypi_0
rich=13.9.4=pypi_0
rpds-py=0.23.1=pypi_0
ruff=0.9.10=pypi_0
safehttpx=0.1.6=pypi_0
scipy=1.15.2=pypi_0
semantic-version=2.10.0=pypi_0
setuptools=75.8.0=py310h06a4308_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
starlette=0.46.1=pypi_0
sympy=1.13.3=pypi_0
tinycss2=1.4.0=pypi_0
tk=8.6.14=h39e8969_0
tomlkit=0.12.0=pypi_0
torch=2.0.1+cu118=pypi_0
torchvision=0.15.2+cu118=pypi_0
tornado=6.4.2=pypi_0
tqdm=4.65.0=pypi_0
traitlets=5.14.3=pypi_0
triton=2.0.0=pypi_0
ttkthemes=3.2.2=pypi_0
typer=0.15.2=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025.1=pypi_0
uc-micro-py=1.0.3=pypi_0
urllib3=2.3.0=pypi_0
uvicorn=0.34.0=pypi_0
wcwidth=0.2.13=pypi_0
webencodings=0.5.1=pypi_0
websockets=11.0.3=pypi_0
wheel=0.45.1=py310h06a4308_0
xz=5.6.4=h5eee18b_1
yarg=0.1.9=pypi_0
yarl=1.18.3=pypi_0
zlib=1.2.13=h5eee18b_1
